fix: move MarketList to new arch
