+++ apps/ledger-live-mobile/src/components/RootNavigator/types/WalletTabNavigator.ts
@@ -1,6 +1,6 @@
 import { NavigatorScreenParams } from "@react-navigation/native";
 import { ScreenName, NavigatorName } from "~/const";
-import { MarketNavigatorStackParamList } from "./MarketNavigator";
+import { MarketNavigatorStackParamList } from "LLM/features/Market/Navigator";
 
 export type WalletTabNavigatorStackParamList = {
   [ScreenName.Portfolio]: undefined;
