+++ apps/ledger-live-mobile/src/newArch/features/Market/components/MarketDataProviderWrapper/index.tsx
