+++ apps/ledger-live-mobile/src/newArch/features/Market/screens/MarketDetail/components/MarketGraph/index.tsx
