+++ apps/ledger-live-mobile/src/newArch/features/Market/screens/MarketList/components/EmptyStarredCoins/index.tsx
